Add files via upload
Was going to show you, you can push it into the bed so the bottom of the rail guides are touching the print surface like this.  But I would recommend turning on supports instead.  I also recommend a toolholder with grooves for your allen keys, or they will be flopping around while printing.  Something like this one... https://www.thingiverse.com/thing:4341801
It is the closest one to the one you showed on your post, with grooves for the keys.  There are hundreds of tool holders and I couldn't find the one you posted.